requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
django-celery==3.1.17
djangorestframework==3.14.0
kombu==5.3.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
vine==5.0.0
wcwidth==0.2.6
